+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/RegistrationRestRepositoryIT.java
@@ -7,19 +7,24 @@
  */
 package org.dspace.app.rest;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.Iterator;
 import java.util.List;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.commons.lang3.StringUtils;
 import org.dspace.app.rest.matcher.RegistrationMatcher;
 import org.dspace.app.rest.model.RegistrationRest;
 import org.dspace.app.rest.test.AbstractControllerIntegrationTest;
 import org.dspace.eperson.RegistrationData;
 import org.dspace.eperson.dao.RegistrationDataDAO;
+import org.dspace.services.ConfigurationService;
 import org.hamcrest.Matchers;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +33,9 @@ public class RegistrationRestRepositoryIT extends AbstractControllerIntegrationT
 
     @Autowired
     private RegistrationDataDAO registrationDataDAO;
+
+    @Autowired
+    private ConfigurationService configurationService;
 
     @Test
     public void findByTokenTestExistingUserTest() throws Exception {
@@ -40,6 +48,8 @@ public class RegistrationRestRepositoryIT extends AbstractControllerIntegrationT
                    .andExpect(status().isOk())
                    .andExpect(
                        jsonPath("$", Matchers.is(RegistrationMatcher.matchRegistration(email, eperson.getID()))));
+
+        registrationDataDAO.delete(context, registrationData);
 
         email = "newUser@testnewuser.com";
         createTokenForEmail(email);
@@ -87,4 +97,75 @@ public class RegistrationRestRepositoryIT extends AbstractControllerIntegrationT
                                 .contentType(contentType))
                    .andExpect(status().isCreated());
     }
+
+    @Test
+    public void registrationFlowTest() throws Exception {
+        List<RegistrationData> registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertEquals(0, registrationDataList.size());
+
+        ObjectMapper mapper = new ObjectMapper();
+        RegistrationRest registrationRest = new RegistrationRest();
+        registrationRest.setEmail(eperson.getEmail());
+        getClient().perform(post("/api/eperson/registrations")
+                                .content(mapper.writeValueAsBytes(registrationRest))
+                                .contentType(contentType))
+                   .andExpect(status().isCreated());
+        registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertEquals(1, registrationDataList.size());
+        assertTrue(StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), eperson.getEmail()));
+
+        String newEmail = "newEPersonTest@gmail.com";
+        registrationRest.setEmail(newEmail);
+        getClient().perform(post("/api/eperson/registrations")
+                                .content(mapper.writeValueAsBytes(registrationRest))
+                                .contentType(contentType))
+                   .andExpect(status().isCreated());
+        registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertTrue(registrationDataList.size() == 2);
+        assertTrue(StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), newEmail) ||
+                       StringUtils.equalsIgnoreCase(registrationDataList.get(1).getEmail(), newEmail));
+        configurationService.setProperty("user.registration", false);
+
+        newEmail = "newEPersonTestTwo@gmail.com";
+        registrationRest.setEmail(newEmail);
+        getClient().perform(post("/api/eperson/registrations")
+                                .content(mapper.writeValueAsBytes(registrationRest))
+                                .contentType(contentType))
+                   .andExpect(status().is(401));
+
+        assertEquals(2, registrationDataList.size());
+        assertTrue(!StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), newEmail) &&
+                       !StringUtils.equalsIgnoreCase(registrationDataList.get(1).getEmail(), newEmail));
+
+        Iterator<RegistrationData> iterator = registrationDataList.iterator();
+        while (iterator.hasNext()) {
+            RegistrationData registrationData = iterator.next();
+            registrationDataDAO.delete(context, registrationData);
+        }
+    }
+
+    @Test
+    public void forgotPasswordTest() throws Exception {
+        configurationService.setProperty("user.registration", false);
+
+        List<RegistrationData> registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertEquals(0, registrationDataList.size());
+
+        ObjectMapper mapper = new ObjectMapper();
+        RegistrationRest registrationRest = new RegistrationRest();
+        registrationRest.setEmail(eperson.getEmail());
+        getClient().perform(post("/api/eperson/registrations")
+                                .content(mapper.writeValueAsBytes(registrationRest))
+                                .contentType(contentType))
+                   .andExpect(status().isCreated());
+        registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertEquals(1, registrationDataList.size());
+        assertTrue(StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), eperson.getEmail()));
+        Iterator<RegistrationData> iterator = registrationDataList.iterator();
+        while (iterator.hasNext()) {
+            RegistrationData registrationData = iterator.next();
+            registrationDataDAO.delete(context, registrationData);
+        }
+    }
+
 }
